event update yazıldı

+++ app/Services/EventService.php
@@ -27,10 +27,24 @@ class EventService extends BaseService
         return $this->eventRepository->create($data);
     }
 
+    public function update(string $id,array $data)
+    {
+        if (isset($data['photo'])){
+            $event = $this->find($id);
+            /*if ($event && $event->photo){
+                $this->replaceFile($data['photo'],'events',$event->photo ?? '');
+            }else{
+                $this->uploadFile($data['photo'],'events');
+            }*/
+            $this->replaceFile($data['photo'],'events',$event->photo ?? '');
+
+        }
+        $this->eventRepository->update($id,$data);
+    }
     public function delete(string $id) : bool
     {
         $event = $this->find($id);
-        if ($event){
+        if ($event && $event->photo){
             $this->deleteFile($event->photo);
         }
      return $this->eventRepository->delete($id);
